V3 Update is out
Now it is organised into different folders for the commands. The memes take from dankmemes instead of memes. The osu command is the only thing needing fixing however I moved it to the fun catagory. G stats is now G character in commands however file name will stay the same.